Added v_ptrtov/vptrtov(T) to vector and vector(T) (see body)
ptrtov takes an existing pointer and adapts it for use with vector.
No allocation/copying takes place at all -- this is simply an assignment;
if a pointer was dynamically allocated clientside, and the client then
decides to use that pointer as a vector, ptrtov will help them do just that.

The programmer must be very careful to make sure that
special attention is taken with respect to memory management --
since the pointer is a preexisting resource, and is not exclusively
owned by the vector -- be sure not to manipulate the "naked"
pointer when the vector is using it...or undefined behavior
may result. When in doubt, just do a deep copy of the pointer
instead and save yourself the trouble.