New idea - creating pre-instantiated typed containers using the template headers/sources. (see body)
At first, I figured I'd use the pre-processor with the -E -CC option (-CC keeps comments)
but there would be a lot to sift through in terms of non-relevant code -- and this whole
thing is still time consuming, regardless. There are virtues to having
typed containers that are ready to go that do not require the somewhat bizarre method
of "template instantiation" by including source files in a preprocessor directive to
replace a type parameter.

All I needed was my template header and template source, and a simple find-and-replace.
I've gotten vector_char_ptr done now, and vector_char_ptr will be the basis for which
the other types will be created (char_ptr doesn't clash with anything else, unlike str,
which may trigger portions of the keyword 'struct' when doing a find-and-replace).

Coming up:
Once list for (void *) is done, I will create the template header/source for list,
and the template instantiations (using the preprocessor "hack", as in
vector_int.h, vector_float.h, vector_str.h, and their source files) --
then, I will create the same pre-instantiated typed list containers.

This may or may not work out in the long run...
What if I want to make a change? The preprocessor template header/source pair isn't
hard to change -- but to change each of the pre-instantiated template files
just to fix a single function?  hmm...