Typed vectors are now complete (see body)
The header/source pairs of
vector_float, vector_int, and vector_str
are in the temp folder, and will not compile during use of Make.
Instead, the built in types have been provided as typed containers --
this is particularly useful when you need a container that will
be part of a more complex object -- and need to get as much
performance as possible. The (void *)-based containers are great,
as they can hold anything, but suffer from worse performance
than that of a typed container.

The templated containers are a good solution, but can be
cumbersome to instantiate (though once prepared,
it is no big deal) -- the only issue with the 'discrete'
typed containers is that if a bug or problem is found...
all of the files will have to be modified, which will be
very tedious. Having one pair of files for virtually
any type from the present and future is very good
from a maintenance point of view, but we will see
how this works out in the long run.
C does not support generics or templates natively,
and it seems that natively typed data structures
are really the way to go if one wants to exploit
as much performance as possible.