See folder test__vptr - simplified vector ADT for pointer types (typename *) (see body)
After doing CS214 Systems Programming at Rutgers in Fall 2019,
I realized that the data structures I wrote in gcslib were a bit bloated.
I myself opted not to use even my original vector data structure variations,
which were my most "mature" ADTs out of the API.

For my lab assignments, I would take the best parts of vector and simplify it --
take out the excess code and remove what I didn't need.
(like using struct typetable and struct iterator).
I noticed that most of the time, I was dealing with pointers to char (char *),
or pointers to structs (struct foobar *) -- so this version of vector,
called vptr_t, is a vector ADT designed for pointer types.

It does not do any memory management of its elements
(no deep copies, and no deallocations/destruction of elements --
that is the left up to the client to do, and I've provided a test.c source file
that demonstrates how one would go about doing this.)

It also does not require the typetable,
because each element will be the size of a pointer -- and if a user
wants to do a search/sort (and needs a comparator) -- it is not hard to write one
on your own. Likewise, an iterator is also provided with vptr_t -- it does not
employ the use a virtual function table like before (my plan was to recreate the C++ STL's Algorithm library,
and that required that all containers conform to the same iterator interface...but it is a pain to use in practice,
since we do not have operator overloading in C -- as far as notation, it is quite messy.)

vptr_t's iterator is simply a (void **), typedefed as a vptr_iter_t.
Pointer arithmetic may be performed on it as expected, so it will at least
"feel" like a std::vector<_Tp>::iterator.

The test.c source file includes two common use cases for vptr_t --
one test function demonstrates storage of dynamically allocated (char *),
and another test function demonstrates storage of int.
Using vptr_t to store int is a bit wasteful, as you do not need sizeof(void *)
bytes per element to store an int -- but at least vptr_t is flexible in this manner.

I've also decided to make vptr_t a non-opaque type. (it can be allocated on the stack and the heap).
vptr_new and vptr_delete macros are provided if heap allocation is still preferred.

If any user decides to access fields directly and not use the functions provided,
they do so at their peril -- treat vptr_t as if it were opaque (even through it is not).